Add spreadsheet_row_results table
The spreadsheet_row_results table should either store the caclulated
column values for each calculation or an adjusted column value.

+++ db/migrate/10_add_inheritable_column_to_calculations.rb
@@ -20,10 +20,10 @@
 
 class AddInheritableColumnToCalculations < ActiveRecord::Migration[4.2]
   def self.up
-    add_column :calculations, :inheritable, :boolean, default: false
+    add_column :calculations, :inheritable, :boolean, default: false if table_exists?(:calculations)
   end
 
   def self.down
-    remove_column :calculations, :inheritable
+    remove_column :calculations, :inheritable if table_exists?(:calculations)
   end
 end
